Created db.js and routes.js
