+++ server.js
@@ -1,37 +1,19 @@
 const express = require('express');
 const app = express();
-const dotenv = require('dotenv').config();
 const mongoose = require('mongoose');
-const connection = process.env.URI;
+const app_routes = require('./routes')
 
-let connectionStatus = 'disconnected';
+const{startDatabase}=require('./db')
 
-const startDatabase = async () => {
-    try {
-        await mongoose.connect(connection);
-        connectionStatus = "Congratulations the database has been connected!!";
-    } catch (err) {
-        console.error("Failed to connect to database");
-        connectionStatus = "Failed to connect to database";
-    }
-};
 
-const stopDatabase = async () => {
-    await mongoose.disconnect();
-    connectionStatus = "Database disconnected";
-};
 
-app.get('/', (req, res) => {
-    res.send(connectionStatus);
-});
+app.use('/', app_routes);
 
-app.get("/ping", (req, res) => {
-    res.send('Hello');
-});
 
 app.listen(3000, () => {
     startDatabase();
-    console.log('Port 3000');
+    console.log('Server running on port 3000');
+    
 });
 
 module.exports = app;
